% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKaplanMeier.R
\name{ggKM.legend.extra}
\alias{ggKM.legend.extra}
\title{Add extra label to a ggKM legend}
\usage{
ggKM.legend.extra(input, label)
}
\arguments{
\item{input}{A \code{ggKM} plot object.}

\item{label}{String to append as an extra legend label.}
}
\value{
A deep copy of the input \code{ggKM} object with the added legend label.
}
\description{
Adds an extra text label to an existing legend of a \code{ggKM} plot.
}
