% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKaplanMeier.R
\name{ggKM}
\alias{ggKM}
\title{Kaplan–Meier plot}
\usage{
ggKM(
  time,
  status,
  group = NULL,
  breaks.s = seq(0, 1, 0.25),
  breaks.t = NULL,
  CI = "cloglog",
  CI.alpha = 0.2,
  colors = (ggsci::pal_nejm())(8),
  grid.color = grDevices::rgb(0.95, 0.95, 0.95),
  grid.s = seq(0, 1, 0.25),
  grid.t = NULL,
  grid.width = 0.5,
  legend.direction = "vertical",
  legend.just = "center",
  legend.label.pos = "left",
  legend.labels = NULL,
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.pos = c(0.9, 0.9),
  legend.text.align = 1,
  line.width = 0.5,
  line.height = 0.025,
  risk.table = TRUE,
  risk.table.margin = 16,
  risk.table.prop = 0.2,
  textsize.axis = 12,
  textsize.legend = 12,
  textsize.risk = 12,
  title.s = "Survival",
  title.t = "Time"
)
}
\arguments{
\item{time}{Numeric vector of follow-up times.}

\item{status}{Numeric event indicator (\code{1} = event, \code{0} = censored).}

\item{group}{Optional integer vector grouping variable.}

\item{breaks.s}{Y-axis (survival) tick marks. Default = \code{seq(0, 1, 0.25)}.}

\item{breaks.t}{X-axis (time) tick marks. Default =
\code{seq(0, max(time), by = 12)}.}

\item{CI}{String indicating the CI type:
\itemize{
\item \code{"none"}: none
\item Pointwise confidence intervals:
\itemize{
\item \code{"cloglog"} (default): pointwise CI using Greenwood's variance¹
with complementary log–log transformation²
\item \code{"modcloglog"}: \code{"cloglog"} with the lower confidence limit
modified according to the effective sample size at each censored
observation³
\item \code{"Rothman"}: pointwise CI using Rothman's binomial method⁴ (via
the \code{WHKMconf} package)
\item \code{"TG"}: pointwise CI using the Thomas–Grunkemeier
likelihood-ratio method⁵ (via the \code{WHKMconf} package)
\item \code{"BPCP"}: pointwise CI using the beta product confidence
procedure⁶ (via the \code{bpcp} package)
}
\item Simultaneous confidence bands:
\itemize{
\item \code{"Nair"}: simultaneous confidence bands using Nair's
log-transformed equal precision method⁷ (via the \code{WHKMconf} package)
\item \code{"HM"}: simultaneous confidence bands using the
Hollander–McKeague likelihood-ratio method⁸ (via the \code{WHKMconf} package)
}
}}

\item{CI.alpha}{Alpha transparency of the confidence intervals. Default =
\code{0.2}.}

\item{colors}{Vector of colors of the survival curves. Default =
\code{ggsci::pal_nejm()(8)}.}

\item{grid.color}{Gridline color. Default = \code{grDevices::rgb(0.95, 0.95, 0.95)}.}

\item{grid.s}{Horizontal gridline positions. Default = \code{seq(0, 1, 0.25)}.}

\item{grid.t}{Vertical gridline positions. Default = \code{NULL}.}

\item{grid.width}{Gridline thickness. Default = \code{0.5}.}

\item{legend.direction}{Legend orientation; either \code{"vertical"} (default) or
\code{"horizontal"}.}

\item{legend.just}{Alignment anchor for the legend relative to its position.
Can be a single keyword (e.g., \code{"center"}), a keyword pair specifying
horizontal and vertical justification (e.g., \code{c("left", "top")},
\code{c("right", "bottom")}) respectively, or a numeric vector of length 2 giving
relative coordinates within the plot area. Default = \code{"center"}.}

\item{legend.label.pos}{Legend label position (\code{"left"} or \code{"right"})
relative to the legend symbol. Default = \code{"left"}.}

\item{legend.labels}{Character vector of group labels.}

\item{legend.ncol}{Integer specifying the number of columns in the legend.
Default = \code{NULL}.}

\item{legend.nrow}{Integer specifying the number of rows in the legend.
Default = \code{NULL}.}

\item{legend.pos}{Position of the legend. Can be a keyword such as \code{"none"},
\code{"left"}, \code{"right"}, \code{"bottom"}, or \code{"top"}, or a numeric vector of length 2
giving relative coordinates within the plot area. Default = \code{c(0.9, 0.9)}.
Set to \code{"none"} if \code{group} is \code{NULL}.}

\item{legend.text.align}{Legend text alignment: \code{0} = left, \code{0.5} = center,
\code{1} = right. Default = \code{1}.}

\item{line.width}{Line width survival curves and censor marks. Default = \code{0.5}.}

\item{line.height}{Height of censor marks. Default = \code{0.025}.}

\item{risk.table}{Logical; if \code{TRUE}, show risk table. Default = \code{TRUE}.}

\item{risk.table.margin}{Numeric value specifying the horizontal spacing
between the risk table labels and the risk table. Default = \code{16}.}

\item{risk.table.prop}{Relative height of the risk table. Default = \code{0.2}.}

\item{textsize.axis}{Axis text size. Default = \code{12}.}

\item{textsize.legend}{Legend text size. Default = \code{12}.}

\item{textsize.risk}{Risk table text size. Default = \code{12}.}

\item{title.s}{Y-axis (survival) title. Default = \code{"Survival"}.}

\item{title.t}{X-axis (time) title. Default = \code{"Time"}.}
}
\value{
A \code{ggplot2} object (or \code{patchwork} composite if \code{risk.table = TRUE}).
}
\description{
Generates a Kaplan–Meier plot with optional confidence intervals and risk
table.
}
\examples{
data <- survival::lung
g <- ggKM(data$time * 12 / 365.2425, data$status - 1, data$sex, 
          breaks.t = seq(0, 30, 6), legend.labels = c("Male", "Female"),
          title.s = "Overall survival", title.t = "Time (months)")
print(g)
}
\references{
\enumerate{
\item Greenwood, M., 1926. A report on the natural duration of cancer. In:
\emph{Reports on Public Health and Medical Subjects}, 33, pp. 1–26. London:
Her Majesty’s Stationery Office, Ministry of Health.
\item Klein, J.P., Logan, B., Harhoff, M. and Andersen, P.K., 2007. Analyzing
survival curves at a fixed point in time. \emph{Statistics in Medicine}, 26(24),
pp. 4505–4519.
\item Dorey, F.J. and Korn, E.L., 1987. Effective sample sizes for confidence
intervals for survival probabilities. \emph{Statistics in Medicine}, 6(6), pp.
679–687.
\item Rothman, K.J., 1978. Estimation of confidence limits for the cumulative
probability of survival in life table analysis. \emph{Journal of Chronic Diseases},
31(8), pp. 557–560.
\item Thomas, D.R. and Grunkemeier, G.L., 1975. Confidence interval estimation
of survival probabilities for censored data. \emph{Journal of the American
Statistical Association}, 70(352), pp. 865–871.
\item Fay, M.P., Brittain, E.H. and Proschan, M.A., 2013. Pointwise confidence
intervals for a survival distribution with small samples or heavy censoring.
\emph{Biostatistics}, 14(4), pp. 723–736.
\item Nair, V.N., 1984. Conﬁdence bands for survival functions with censored
data: a comparative study. \emph{Technometrics}, 26, pp. 265–275.
\item Hollander, M. and McKeague, I.W., 1997. Likelihood ratio-based confidence
bands for survival functions. \emph{Journal of the American Statistical
Association}, 92(437), pp. 215–226.
}
}
